updated Navbar component
