--- food-hut/styles/index.module.css
@@ -1,38 +1,0 @@
-.header {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 1rem 1.5rem;
-}
-
-.menu {
-  /* display: none; */
-}
-
-.links {
-  display: none;
-  align-items: center;
-  justify-content: center;
-  gap: 2rem;
-}
-
-.links p {
-  cursor: pointer;
-  font-size: 1.1rem;
-}
-
-.links p:last-child {
-  background: #f54748;
-  padding: 0.7rem 0.9rem;
-  border-radius: 100px;
-}
-
-@media (width > 980px) {
-  .menu {
-    display: none;
-  }
-
-  .links {
-    display: flex;
-  }
-}
